move shell_clock include data in c file
